refactor(moonrun): separate host state from v8
